Remove my setting the python enviroment for coc
The default behavior now is to set a python interpreter per project,
using :CocCommand, then python.setInterpreter, and finally choosing the
one that you want. You have coc-status in your status line, so you see
which one is activated. By default it is the first one in your $PATH.

Maybe this is what you want:
https://github.com/neoclide/coc-python/tree/master/src/interpreter/autoSelection.

But in the github issues is not marked as the answer, so here you can find one
solution when using conda:

Link: https://github.com/neoclide/coc-python/issues/55#issuecomment-617499047
Workaround: If conda is not activated, it activates it and sets it as
the default python.

        In your init.vim/ file, add the following:

        if $CONDA_PREFIX == ""
          let s:current_python_path=$CONDA_PYTHON_EXE
        else
          let s:current_python_path=$CONDA_PREFIX.'/bin/python'
        endif
        call coc#config('python', {'pythonPath': s:current_python_path})